Refactor domain modules "services" abstract to `index` file
The services present public interface of the given domain module,
index file is natural place for it
