+++ api/src/domain/access/account/account.repository.ts
@@ -1,8 +1,8 @@
 import { Account } from './account';
 
-export type AccountRepository = {
+export interface AccountRepository {
   getByEmail(email: string): Promise<Account>;
   create(email: string, passwordHash: string): Promise<Account>;
   saveAccount(account: Account): Promise<Account>;
   getById(id: string): Promise<Account>;
-};
+}
