+++ api/src/domain/access/index.ts
@@ -24,3 +24,19 @@ export async function loginWithPassword(
 
   throw 'invalid login data';
 }
+
+export type registerAccountWithPasswordDI = {
+  accountRepository: AccountRepository;
+  encryptor: PasswordEncryptor;
+};
+
+export async function registerAccountWithPassword(
+  di: registerAccountWithPasswordDI,
+  email: string,
+  password: string
+) {
+  const { accountRepository, encryptor } = di;
+  const passwordHash = await encryptor.hashPassword(password);
+  const account = await accountRepository.create(email, passwordHash);
+  return accountRepository.saveAccount(account);
+}
