+++ api/src/domain/services/create-character-for-account.logic.ts
@@ -1,8 +1,5 @@
 import { AccountRepository } from '@/domain/access/account/account.repository';
-import {
-  createCharacter,
-  createCharacterDI,
-} from '@/domain/game/services/create-character.logic';
+import { createCharacter, createCharacterDI } from '@/domain/game';
 
 export type createCharacterForAccountDI<TRaceKey> = {
   accountRepository: AccountRepository;
@@ -21,11 +18,12 @@ export async function createCharacterForAccount<TRaceKey>(
 
   await accountRepository.saveAccount({
     ...account,
-    // MARCHW: do we really need to check if the list will be unique?
+    // MARCHW: I dropped here uniq check that u have before.
+    // do we really need to check if the list will be unique?
     // IMO not. the api should throw an error in such case instead of quitely ignore it.
     // and the error should go from database unique constrain that
     // should be at accountId+characterId columns
-    characters: [...new Set([...account.characters, character.id])],
+    characters: [...account.characters, character],
   });
 
   return character;
