--- api/src/server/db/entities/access/account-character.ts
@@ -1,15 +1,0 @@
-import { Entity, PrimaryColumn, ManyToOne } from 'typeorm';
-
-import { Account } from './account';
-
-@Entity({ name: 'access_accounts_characters' })
-export class AccountCharacter {
-  @PrimaryColumn('uuid')
-  public id: string;
-
-  @ManyToOne(
-    () => Account,
-    account => account.characters
-  )
-  account: Account;
-}
